refactor(conditor): Move getRnsr from affAlign.js to rnsr.js
So that it can be used by other statements.
